Add a new validate function to remove the need to a custom vpnor protocol implimentation.
