+++ flash.c
@@ -54,7 +54,7 @@ static struct backend flash_mtd_backed = {
 	.erase = flash_erase,
 	.write = flash_write,
 	.lpc_reset = lpc_reset,
-	.protocol_negotiate_version = protocol_negotiate_version_default,
+	.validate = NULL,
 	.flash_bmap = NULL,
 	.erase_size_shift = 0,
 	.block_size_shift = 0,
