+++ protocol_negotiate_version.c
@@ -37,7 +37,7 @@ static const struct protocol_ops *protocol_ops_map[] = {
 	[2] = &protocol_ops_v2,
 };
 
-int protocol_negotiate_version_default(struct mbox_context *context,
+int protocol_negotiate_version(struct mbox_context *context,
 				   uint8_t requested)
 {
 	/* Check we support the version requested */
